channel lock: fix priority order
The intent of this line was to produce a vector containing the index of
each number in priority_order:

    np.arange(len(layout.channels))[priority_order]

but this does nothing; the result is the same as priority_order.

In this case the spec and the original intent was correct; this
behaviour makes no sense. The existing tests aren't too bad, we were
just unlucky in that the order in 4+5+0 used for the tests happened to
be correct.